# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  category: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  stock: number;
  imageUrl: string;
  featured: boolean;
  rating: number;
  reviewCount: number;
  createdAt?: string;
}

export interface ProductInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  description: string;
  /** @minLength 2 */
  category: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 0
     * @nullable
     */
  compareAtPrice?: number | null;
  /** @minimum 0 */
  stock: number;
  imageUrl: string;
  featured?: boolean;
}

/**
 * Product fields to update
 */
export type ProductUpdate = ProductInput;

export interface OrderItemInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface ShippingQuoteInput {
  /** @minLength 8 */
  postalCode: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface ShippingOption {
  id: string;
  carrier: string;
  service: string;
  price: number;
  deliveryDays: number;
  description: string;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  status: string;
  subtotal: number;
  shipping: number;
  total: number;
  shippingOption: ShippingOption;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 2 */
  customerName: string;
  customerEmail: string;
  /** @minLength 8 */
  postalCode: string;
  shippingOptionId: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface Profile {
  id: number;
  name: string;
  email: string;
  phone: string;
  postalCode: string;
  city: string;
  state: string;
}

export interface ProfileUpdate {
  /** @minLength 2 */
  name: string;
  email: string;
  phone: string;
  postalCode: string;
  city: string;
  state: string;
}

export interface SalesPoint {
  label: string;
  value: number;
}

export interface TopProduct {
  name: string;
  sales: number;
  revenue: number;
}

export interface RecentOrder {
  id: number;
  customerName: string;
  total: number;
  status: string;
  createdAt: string;
}

export interface DashboardSummary {
  revenue: number;
  revenueChange: number;
  orders: number;
  ordersChange: number;
  customers: number;
  products: number;
  inventoryUnits: number;
  lowStock: number;
  salesByDay: SalesPoint[];
  topProducts: TopProduct[];
  recentOrders: RecentOrder[];
}

export type AdminRole = typeof AdminRole[keyof typeof AdminRole];


export const AdminRole = {
  admin: 'admin',
  operator: 'operator',
} as const;

export interface AdminLoginInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface AdminSetupStatus {
  needsSetup: boolean;
}

export interface AdminSetupInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 12 */
  password: string;
  /** @minLength 8 */
  setupSecret: string;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: AdminRole;
  active: boolean;
  /** @nullable */
  lastLoginAt: string | null;
  createdAt: string;
}

export interface AdminUserInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  role: AdminRole;
}

export interface AdminUserUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 3 */
  email?: string;
  /** @minLength 8 */
  password?: string;
  role?: AdminRole;
  active?: boolean;
}

export interface AdminSession {
  user: AdminUser;
}

export type ListProductsParams = {
search?: string;
category?: string;
featured?: boolean;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListOrdersParams = {
email?: string;
};