# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Sign in to the store administration area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Authenticated administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
        "429":
          description: Too many login attempts
...
  /admin/auth/setup:
    post:
      operationId: adminSetup
      tags: [admin-auth]
      summary: Configure the first administrator with an environment-controlled secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSetupInput"
      responses:
        "201":
          description: First administrator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "400":
          description: Invalid setup data
        "401":
          description: Invalid setup secret
        "409":
          description: An administrator is already configured
        "429":
          description: Too many setup attempts
...
  /admin/auth/me:
    get:
      operationId: getAdminSession
      tags: [admin-auth]
      summary: Get the current administration session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Authentication required
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin-users]
      summary: List store team users
      responses:
        "200":
          description: Store team users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Authentication required
        "403":
          description: Administrator role required
    post:
      operationId: createAdminUser
      tags: [admin-users]
      summary: Create a store team user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Invalid user
        "401":
          description: Authentication required
        "403":
          description: Administrator role required
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin-users]
      summary: Update a store team user
      parameters:
        - $ref: "#/components/parameters/AdminUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Invalid user
        "401":
          description: Authentication required
        "403":
          description: Administrator role required
        "404":
          description: User not found
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [orders]
      summary: List all orders for the store team
      responses:
        "200":
          description: Store orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
components:
  parameters:
...
    AdminUserId:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
  schemas:
...
    ShippingOption:
      type: object
      required: [id, carrier, service, price, deliveryDays, description]
      properties:
        id: { type: string }
        carrier: { type: string }
        service: { type: string }
        price: { type: number }
        deliveryDays: { type: number }
        description: { type: string }
...
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId: { type: number }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        total: { type: number }
    Order:
      type: object
      required: [id, customerName, customerEmail, status, subtotal, shipping, total, shippingOption, items, createdAt]
      properties:
        id: { type: number }
        customerName: { type: string }
        customerEmail: { type: string }
        status: { type: string }
        subtotal: { type: number }
        shipping: { type: number }
        total: { type: number }
        shippingOption: { $ref: "#/components/schemas/ShippingOption" }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt: { type: string, format: date-time }
...
    AdminRole:
      type: string
      enum: [admin, operator]
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 8 }
...
    AdminSetupInput:
      type: object
      required: [name, email, password, setupSecret]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 12 }
        setupSecret: { type: string, minLength: 8 }
    AdminUser:
      type: object
      required: [id, name, email, role, active, lastLoginAt, createdAt]
      properties:
        id: { type: number }
        name: { type: string }
        email: { type: string }
        role:
          $ref: "#/components/schemas/AdminRole"
        active: { type: boolean }
        lastLoginAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    AdminUserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 8 }
        role:
          $ref: "#/components/schemas/AdminRole"
    AdminUserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 8 }
        role:
          $ref: "#/components/schemas/AdminRole"
        active: { type: boolean }
    AdminSession:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"