# lib/api-zod/src/generated/types/adminRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type AdminRole = typeof AdminRole[keyof typeof AdminRole];


export const AdminRole = {
  admin: 'admin',
  operator: 'operator',
} as const;