# lib/api-zod/src/generated/types/adminSetupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSetupInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 12 */
  password: string;
  /** @minLength 8 */
  setupSecret: string;
}