# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: AdminRole;
  active: boolean;
  /** @nullable */
  lastLoginAt: Date | null;
  createdAt: Date;
}