# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface AdminUserInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  role: AdminRole;
}